svg_loader: css style functions moved to a separate file
